src/APIGateway/requirements.txt
Flask==2.3.2
Werkzeug==2.3.6
PyJWT==2.7.0
cryptography==41.0.3
requests==2.31.0
grpcio
grpcio-tools
kafka-python==2.0.2
protobuf==6.30.0
